api/requirements.txt
flask~=3.1.0
python-dotenv~=1.0.1
xrpl-py~=3.0.0